+++ src/app/login/recover-password/recovery-new-password/components/RecoverPassword/RecoverPassword.tsx
@@ -1,0 +1,51 @@
+"use client";
+import DynamicForm from "@/app/components/DynamicForm/DynamicForm";
+import logoDesktop from "@/assets/images/Walpapers/Wallpaper-1.png";
+import logoMobile from "@/assets/images/Walpapers/wallpaper-mobile-rp.png"
+import Image from "next/image";
+import { nesPasswordStyles } from "./styles";
+import useChangePassword,{changePasswordFields} from "../../../hooks/useChangePassword/useChangePassword";
+
+const RecoverPassword = () => {
+  const { handleChange, isLoading } = useChangePassword();
+
+  return (
+    <div className={nesPasswordStyles.container}>
+      {/* Imagen de fondo */}
+      <Image
+        src={logoDesktop}
+        alt="Fondo DR Security (desktop)"
+        fill
+        priority
+        className={`${nesPasswordStyles.image} hidden sm:block`}
+      />
+      <Image
+        src={logoMobile}
+        alt="Fondo DR Security (móvil)"
+        fill
+        priority
+        className={`${nesPasswordStyles.image} block sm:hidden`}
+      />
+
+      {/* Overlay azul */}
+      <div className={nesPasswordStyles.bgOverlay} />
+
+      {/* Contenido centrado */}
+      <div className={nesPasswordStyles.contentCenter}>
+        <div className={nesPasswordStyles.card}>
+          <DynamicForm
+            title=""
+            fields={changePasswordFields}
+            submitLabel="Guardar"
+            onSubmit={handleChange}
+            loading={isLoading}
+          />
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default RecoverPassword;
+
+